+++ resources/views/mail/contact-reply.blade.php
@@ -1,0 +1,1 @@
+<p>Hello {{ $contactMessage->full_name }},</p><p>{!! nl2br(e($replyBody)) !!}</p>